remove call to deprecated POSIX::tmpnam (perl 26)
This patch replaces the call to POSIX::tmpnam with a call to
File::Temp.

File::Temp object handles file cleanup when the object is destroyed, so
the call to unlink($filename) is no longer needed.

While I was at it, I've replaced a call to POSIX::_exit with good old
die, so frozen-bubble no longer depends on POSIX.

This patch also fixes debian bug #866321 [1] and will be shortly applied
to Debian version of frozen-bubble.

Instructions to test this patch are shown in [1].

All the best

[1] https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=866321